requirements.txt
Django==1.11.7
django-extensions==1.9.7
djangorestframework==3.7.3
Markdown==2.6.9
psycopg2==2.7.3.2
pytz==2017.3
selenium==3.8.0
six==1.11.0
typing==3.6.2
